requirements.txt
crewai
crewai_tools
load_dotenv
langchain_groq
yt_dlp==2023.11.14 
youtube-transcript-api==0.6.1